Move utils to internal
