+++ config/types_site.go
@@ -5,7 +5,7 @@ import (
 	"log"
 
 	"github.com/creasty/defaults"
-	"github.com/labd/mach-composer/utils"
+	"github.com/labd/mach-composer/internal/utils"
 	"gopkg.in/yaml.v3"
 )
